backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0              # async MongoDB driver
pymongo==4.7.2
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
Pillow==10.3.0
opencv-python-headless==4.9.0.80
anthropic==0.25.8         # Claude API
openai==1.30.1            # optional OpenAI
httpx==0.27.0
python-multipart==0.0.9   # file uploads
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9            # rate limiting
